Scroll a long name in the App Switcher too
