+++ CHANGELOG.md
@@ -60,8 +60,8 @@ Vorssaint 3.3.3 adds full manual and temperature-based Fan Control, directional 
   window's state along the bottom of the picture, the name and its subtitle
   underneath — and run down the screen beside a Dock on the left or right
   instead of across it. A hovered panel drops its header; a pinned one keeps it.
-  A name too long for its band scrolls while the pointer is on that card.
-  Thanks to @PathGao.
+  A name too long for its band scrolls while the pointer is on that card, in the
+  App Switcher as well. Thanks to @PathGao.
 - Fan Control now prepares stopped fans before taking manual control and keeps a
   failed attempt visible instead of silently returning to Automatic.
 - The Clean URL settings fields now take a click anywhere across their row. Their
