+++ Sources/Vorssaint/UI/Switcher/DockPreviewPanelView.swift
@@ -349,84 +349,6 @@ private struct NativeWindowDragHandle: NSViewRepresentable {
     }
 }
 
-/// The window's name, scrolled instead of clipped while the pointer is on the
-/// card. A middle ellipsis is what a long name looks like at rest, and it eats
-/// exactly the part that tells two windows of one app apart -- but the pointer
-/// is already on the card by the time anybody is trying to read it, so that is
-/// where the name is allowed to move.
-private struct DockPreviewTitleLine: View {
-    let text: String
-    let weight: Font.Weight
-    let scrolls: Bool
-
-    @State private var began: Date?
-
-    private static let size: CGFloat = 13
-    private static let speed: CGFloat = 26
-    private static let gap: CGFloat = 20
-    /// Long enough that arriving on a card is a still frame, not a moving one.
-    private static let startDelay: Double = 0.4
-    private static let fade: CGFloat = 6
-
-    private var width: CGFloat { DockPreviewSupport.cardTitleTextWidth }
-
-    private var overflows: Bool {
-        DockPreviewSupport.titleOverflows(text,
-                                          width: width,
-                                          fontSize: Self.size,
-                                          weight: weight == .semibold ? .semibold : .regular)
-    }
-
-    var body: some View {
-        Group {
-            if scrolls, overflows {
-                scrollingLine
-            } else {
-                Text(text)
-                    .lineLimit(1)
-                    .truncationMode(.middle)
-            }
-        }
-        .font(.system(size: Self.size, weight: weight))
-        .frame(width: width, alignment: .leading)
-    }
-
-    /// Two copies a gap apart, slid left and wrapped, so the name reads
-    /// continuously instead of snapping back at the end. Time is counted from
-    /// the moment the pointer arrived, not from the clock: against the clock
-    /// the line appeared at whatever phase the cycle happened to be in, which
-    /// is why it looked like it started from the middle.
-    private var scrollingLine: some View {
-        let measured = DockPreviewSupport.titleWidth(text,
-                                                     fontSize: Self.size,
-                                                     weight: weight == .semibold ? .semibold : .regular)
-        let run = measured + Self.gap
-        let cycle = Double(run) / Double(Self.speed) + Self.startDelay
-        return TimelineView(.animation(minimumInterval: 1.0 / 30.0, paused: false)) { context in
-            let start = began ?? context.date
-            let phase = context.date.timeIntervalSince(start).truncatingRemainder(dividingBy: cycle)
-            let travelled = max(0, phase - Self.startDelay)
-            HStack(spacing: Self.gap) {
-                Text(text).lineLimit(1).fixedSize()
-                Text(text).lineLimit(1).fixedSize()
-            }
-            .offset(x: -CGFloat(travelled) * Self.speed)
-        }
-        .onAppear { began = Date() }
-        .onChange(of: text) { _, _ in began = Date() }
-        .frame(width: width, alignment: .leading)
-        .clipped()
-        .mask(
-            LinearGradient(stops: [
-                .init(color: .clear, location: 0),
-                .init(color: .black, location: Self.fade / width),
-                .init(color: .black, location: 1 - Self.fade / width),
-                .init(color: .clear, location: 1),
-            ], startPoint: .leading, endPoint: .trailing)
-        )
-    }
-}
-
 private struct DockPreviewCard: View {
     let window: SwitcherItem
     let preview: CGImage?
@@ -586,9 +508,10 @@ private struct DockPreviewCard: View {
                 // read at a glance and the selection has to carry; a Dock
                 // preview holds the windows of one app, where the name is the
                 // only thing telling them apart.
-                DockPreviewTitleLine(text: window.displayTitle,
-                                     weight: isSelected ? .semibold : .regular,
-                                     scrolls: isHovering)
+                ScrollingTitle(text: window.displayTitle,
+                               weight: isSelected ? .semibold : .regular,
+                               width: DockPreviewSupport.cardTitleTextWidth,
+                               scrolls: isHovering)
                     .foregroundStyle(.primary)
                 if let subtitle = window.displaySubtitle {
                     Text(subtitle)
